Victory condition, weapon tweaks

+++ Assets/Scripts/Passives/PassivePlayerHealthRegeneration.cs
@@ -2,11 +2,11 @@ using UnityEngine;
 
 public class PassivePlayerHealthRegeneration : Passive
 {
-    public float baseExtraHealthRegenPerSecond = 1.0f;
+    public int baseExtraHealthRegenPerSecond = 1;
 
     public override void ApplyEffect(PlayerStats playerStats)
     {
-        float extraHealthRegenPerSecond = baseExtraHealthRegenPerSecond * level;
+        int extraHealthRegenPerSecond = baseExtraHealthRegenPerSecond * level;
         playerStats.healthRegenRate += extraHealthRegenPerSecond;
     }
     public override void ApplyEffect(Weapon weapon)
